requirements.txt
PyYAML
Jinja2
markdown
requests
beautifulsoup4